+++ src/devpilot_core/modeling/__init__.py
@@ -13,6 +13,7 @@ from devpilot_core.modeling.ollama_adapter import OllamaAdapter
 from devpilot_core.modeling.lmstudio_adapter import LMStudioAdapter
 from devpilot_core.modeling.health import ModelHealthService
 from devpilot_core.modeling.local_provider_health import LocalLlmProviderHealthOptions, LocalLlmProviderHealthReporter
+from devpilot_core.modeling.external_api_pilot import ExternalApiProviderPilotOptions, ExternalApiProviderPilotReporter, FakeExternalApiProvider
 from devpilot_core.modeling.capabilities import CapabilityMatrix
 from devpilot_core.modeling.budget import BudgetLedger
 from devpilot_core.modeling.evals import ModelEvalRunner, ModelEvalRunnerConfig
@@ -34,6 +35,9 @@ __all__ = [
     "ModelHealthService",
     "LocalLlmProviderHealthOptions",
     "LocalLlmProviderHealthReporter",
+    "FakeExternalApiProvider",
+    "ExternalApiProviderPilotReporter",
+    "ExternalApiProviderPilotOptions",
     "ModelEvalRunner",
     "ModelEvalRunnerConfig",
     "MockModelAdapter",